+++ discord_twitter_webhooks/send_webhook.py
@@ -9,17 +9,19 @@ from discord_twitter_webhooks import settings
 
 
 def send_embed_webhook(
-    tweet,
+    tweet_id: int,
     link_list: list[str],
     text: str,
     twitter_card_image: str,
+    avatar_url: str,
+    screen_name: str,
     webhook: str = settings.webhook_url,
 ):
     """Send embed to Discord webhook.
 
     Args:
         avatar (str): Avatar URL
-        tweet ([type]): Tweet object
+        tweet_id (int): Tweet id
         link_list (list[str]): List of links from the tweet
         text (str): Text from the tweet
         webhook (str, optional): Webhook URL. Defaults to environment
@@ -33,7 +35,7 @@ def send_embed_webhook(
 
     if twitter_card_image:
         embed.set_image(url=twitter_card_image)
-        settings.logger.debug(f"Tweet: {text}")
+        settings.logger.debug(f"twitter_card_image: {twitter_card_image}")
 
     # Only add image if there is one
     if len(link_list):
@@ -43,24 +45,22 @@ def send_embed_webhook(
         elif len(link_list) > 1:
             # Send images to twitter-image-collage-maker
             # (e.g https://twitter.lovinator.space/) and get a collage back.
-            response = requests.get(url=settings.collage_maker_url, params={"tweet_id": tweet.id})  # noqa: E501, pylint: disable=line-too-long
+            response = requests.get(url=settings.collage_maker_url, params={"tweet_id": tweet_id})  # noqa: E501, pylint: disable=line-too-long
 
             if response.status_code == 200:
                 json_data = json.loads(response.text)
                 settings.logger.debug(f"JSON data from twitter-image-collage-maker: {json_data}")  # noqa: E501, pylint: disable=line-too-long
                 embed.set_image(url=json_data["url"])
             else:
-                settings.logger.error(f"Failed to get response from {settings.collage_maker_url}")  # noqa: E501, pylint: disable=line-too-long
+                settings.logger.error(f"Got {response.status_code} from {settings.collage_maker_url}")  # noqa: E501, pylint: disable=line-too-long
                 embed.set_image(url=link_list[0])
 
-    avatar_url = tweet.user.profile_image_url_https
-    avatar_url = avatar_url.replace("_normal.jpg", ".jpg")
     settings.logger.debug(f"Avatar URL: {avatar_url}")
 
     embed.set_author(
         icon_url=avatar_url,
-        name=tweet.user.screen_name,
-        url=f"https://twitter.com/i/web/status/{tweet.id}",
+        name=screen_name,
+        url=f"https://twitter.com/i/web/status/{tweet_id}",
     )
 
     # Add embed object to webhook
@@ -68,5 +68,21 @@ def send_embed_webhook(
 
     response = hook.execute()
 
-    settings.logger.info(f"Webhook posted for tweet https://twitter.com/i/web/status/{tweet.id}")  # noqa: E501, pylint: disable=line-too-long
+    settings.logger.info(f"Webhook posted for tweet https://twitter.com/i/web/status/{tweet_id}")  # noqa: E501, pylint: disable=line-too-long
+    settings.logger.debug(f"Webhook response: {response}")
+
+
+def send_normal_webhook(msg: str, webhook: str = settings.webhook_url):
+    """Send normal message to Discord webhook.
+
+    Args:
+        msg (str): Message to send
+        webhook (str, optional): Webhook URL. Defaults to environment
+        variable WEBHOOK_URL.
+    """
+    settings.logger.debug(f"Message: {msg}")
+
+    hook = DiscordWebhook(url=webhook, content=msg)
+
+    response = hook.execute()
     settings.logger.debug(f"Webhook response: {response}")